backend/requirements.txt
flask
flask-cors
mistralai
google-genai
numpy
python-dotenv
